requirements.txt
onnxruntime-gpu
insightface
faiss-cpu
numpy
opencv-python
cryptography
uvicorn
python-multipart